Fix FPCR truncation on Apple arm64
`~off_bits` was 32-bit, so the complement cleared FPCR bits 63:32 on every
call.  Cast to unsigned long long before complementing.  Also note in the
docs that a divide-by-zero trap on Apple silicon reads back as enabled even
though the hardware will not raise it.

+++ Src/Base/AMReX.cpp
@@ -1088,7 +1088,7 @@ namespace {
         if (any(off & FPExcept::overflow)) { off_bits |= __fpcr_trap_overflow ; }
         fenv_t env;
         fegetenv(&env);
-        env.__fpcr &= ~off_bits;
+        env.__fpcr &= ~static_cast<unsigned long long>(off_bits);
         env.__fpcr |= on_bits;
         fesetenv(&env);
 #endif
